+++ lib/presentation/pages/youtube/detail/providers/youtube_detail_resource_provider.dart
